requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
pydantic==2.5.2
streamlit==1.29.0
requests==2.31.0
Pillow==10.0.0
PyJWT==2.8.0
bcrypt==4.0.1
